+++ flashinfer/gemm/kernels/cutile/__init__.py
@@ -1,0 +1,19 @@
+# Copyright (c) 2025 by FlashInfer team.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""
+cuTile GEMM kernel implementations (internal).
+
+Import :func:`flashinfer.gemm.is_cuda_tile_available` for the public availability check.
+Import from ``flashinfer.gemm`` for the public kernel APIs.
+"""
